more CR fixes
