+++ packages/server/modules/gatekeeper/domain/billing.ts
@@ -2,7 +2,11 @@ import {
   WorkspacePlanProductPrices,
   WorkspacePricingProducts
 } from '@/modules/gatekeeperCore/domain/billing'
-import { Workspace, WorkspaceAcl } from '@/modules/workspacesCore/domain/types'
+import {
+  Workspace,
+  WorkspaceSeat,
+  WorkspaceSeatType
+} from '@/modules/workspacesCore/domain/types'
 import {
   Nullable,
   Optional,
@@ -15,6 +19,12 @@ import {
 } from '@speckle/shared'
 import { OverrideProperties } from 'type-fest'
 import { z } from 'zod'
+
+export { WorkspaceSeat, WorkspaceSeatType }
+export {
+  GetWorkspaceRoleAndSeat,
+  GetWorkspaceRolesAndSeats
+} from '@/modules/workspacesCore/domain/operations'
 
 export type GetWorkspacePlan = (args: {
   workspaceId: string
@@ -199,20 +209,6 @@ export type ReconcileSubscriptionData = (args: {
   prorationBehavior: 'always_invoice' | 'create_prorations' | 'none'
 }) => Promise<void>
 
-export const WorkspaceSeatType = <const>{
-  Viewer: 'viewer',
-  Editor: 'editor'
-}
-export type WorkspaceSeatType =
-  (typeof WorkspaceSeatType)[keyof typeof WorkspaceSeatType]
-
-export type WorkspaceSeat = {
-  workspaceId: string
-  userId: string
-  type: WorkspaceSeatType
-  createdAt: Date
-  updatedAt: Date
-}
 // Prices
 export type GetRecurringPrices = () => Promise<
   {
@@ -224,26 +220,3 @@ export type GetRecurringPrices = () => Promise<
 >
 
 export type GetWorkspacePlanProductPrices = () => Promise<WorkspacePlanProductPrices>
-
-export type GetWorkspaceRolesAndSeats = (params: {
-  workspaceId: string
-  userIds?: string[]
-}) => Promise<{
-  [userId: string]: {
-    role: WorkspaceAcl
-    seat: Nullable<WorkspaceSeat>
-    userId: string
-  }
-}>
-
-export type GetWorkspaceRoleAndSeat = (params: {
-  workspaceId: string
-  userId: string
-}) => Promise<
-  | {
-      role: WorkspaceAcl
-      seat: Nullable<WorkspaceSeat>
-      userId: string
-    }
-  | undefined
->
